fix: move out api error handling to own file
